refactor: staticInfo 조회를 Projection 쿼리문으로 조회 개선

+++ database/src/main/java/com/pickcar/dailyreport/domain/StaticInfo.java
@@ -15,14 +15,14 @@ import lombok.Setter;
 @AllArgsConstructor
 public class StaticInfo {
 
-    private Long totalVehicleCount;
+    private Integer totalVehicleCount;
 
-    private Long reservedVehicleCount;
+    private Integer reservedVehicleCount;
 
-    private Long notOperableVehicleCount;
+    private Integer notOperableVehicleCount;
 
-    private Long expectedReturnCount;
+    private Integer expectedReturnCount;
 
-    private Long delayedCount;
+    private Integer delayedCount;
 
 }
